feat(mpp): Try to factorize common configuration
In gradle.properties there is now a property allowing to set the location of libsecp256k containing folder.

It can be updated to reflect proper location of the library.

Removed specific flags in .def file and made a generic configure task for cinterop.

Added secp256k1 as a submodule too