Remove fee estimator from NodeParams
NodeParams is supposed to be a simple container for static configuration parameters, having a fee estimator there was a code smell
and it created serialization headaches.
When we add support for LN's update_fee, we will just create new events/actions for the channel FSM, and we could for example have a coroutine that
calls our fee estimator and update channel states wih the appropriate event.