Remove redundant status from IncomingPayment (#148)
This change simplifies the IncomingPayment object and fixes a potential
issue where the Expired status checks could be inconsistent with the
actual invoice's expiration.

An incoming payment is paid if the `received` attribute is not null,
otherwise this payment is waiting for funds to come (incoming funds
can be rejected for any reasons but that's not up to the object to
decide).

IncomingPaymentHandler also the `received` attribute when checking 
invoices' expiration. Payments for expired invoices that have already been 
paid will always be accepted.