Rework wire serialization
We previously tried to use the same interface for lightning messages,
tlv fields and onion packet contents. This was a bad abstraction resulting
in mandatory but unused inherited fields (tag).

We now provide a clean separation between these distinct cases:

* A collection of common codecs are offered in LightningCodecs
* LightningMessage implements message-level wire serialization
* Tlv implements tlv serialization

An important note is that reading and writing are not symmetric: reading
is a static operation whereas writing operates on an instance.
The interfaces now reflect that reality and use companion objects to
implement the reading part.

Fixes #45