Channel: do check for timed-out HTLCs until explicitly told to do so
When we start, we need to wait a bit before we check for timed-out HTLCs, this will give our peer enough time to re-send messages that we may have missed.
So when we receive a new block, we won't perform this check unless we've already received an explicit ChannelEvent.CheckHtlcTimeout message from our local Peer.