pay-to-open: add min amount + min fee
We now enforce a minimum amount for pay-to-open, and a minimum fee. This
means that for small amounts, the fee may be much higher than the
default 0.1 %.

If the total amount of an incoming payment is below the min pay-to-open
amount, and there is at least one pay-to-open part, the payment will be
failed with an `IncorrectPaymentDetails` error.