add support for simple trusted swap-in
The workflow is as follow:

```
  Phoenix                    ACINQ
     |                         |
     |-----swap_in_request---->|
     |<----swap_in_response----|
     |                         |
     | (wait for on-chain tx)  |
     |                         |
     |<----swap_in_pending---->|
     |                         |
     | (wait for confirmation) |
     |                         |
     |<---swap_in_confirmed----|
     |<-----open_channel-------|
     |                         |
```

Phoenix starts by sending a `SwapInRequest` message, in order to get a
bitcoin address. ACINQ will always return the same address, until a
confirmed transaction has been made to that address at which point it
will be rotated. Address reuse is discouraged but supported for UX
purposes.

As soon as ACINQ detects an incoming unconfirmed transaction, it
will emit a `SwapInPending` message. This message will be sent after
every reconnection, with possibly an updated amount, if the user sends
multiple transactions to the same address.

Once the transaction is confirmed, ACINQ will emit a single
`SwapInConfirmed` message and will start attempting to create a channel
which pushes a certain amount to the user. The `SwapInConfirmed` message
is only sent once, its purpose is to be a notification. The channel
open will be attempted until it succeeds. Note that since #209, the
`OpenChannel` message contains a tlv that indicates the origin of the
channel. In the case of a swap-in, the origin will be the bitcoin
address.

Gotchas:
- `SwapInPending.amount` is in _satoshis_. It indicates the amount paid
  to the bitcoin adress by the user;
- `SwapInConfirmed.amount` is in _millisatoshis_. It indicates the
  amount that will actually be pushed to the user in the newly created
  channel (importantly, fees will be deducted).