Change seed derivation purpose
We want to use a different derivation from the current Phoenix app, to ensure
users cannot shoot themselves in the foot by restoring an incompatible seed.

With that change, a user restoring a seed from an old Phoenix wallet will
end up with a different node_id, so its previous channels will not be closed.

That user can then go to the old Phoenix app to move his funds to the new one.

Fixes #193