Store incoming amounts to payments database (#230)
The `StoreIncomingAmount` action is saved in the payments database as 
an `IncomingPayment` object. The payment database has a new method 
to allow insertion of spontaneous payments, such as swap-ins.

We use the channel origin to set the  incoming payment's `receivedWith`
attribute. If channel origin is not defined, the payment origin is a swap-in.