Precompute channel keys
Key manager will compute all channel keys and will pass generated private keys back to the channel where they will be kept and used.
This is much more efficient than re-computing them from the channel path everytime as we do now.
Another option would be to cache private keys in the key manager (as we do on eclair) but it's not as useful for mobile apps as we are not likely to delegate signing to an external module.