Add WeakRandom prng
n case of catastrophic failures of the SecureRandom instance, we add a
secondary randomness source that we mix into the random stream.

This is a somewhat weak random source and should not be used on its own,
but it doesn't hurt to xor it with the output of SecureRandom.

The `runtimeEntropy` should be customized for iOS and Android.