Store payments parts in incoming payments database
This fixes ian issue where here a channel could replay send an htlc to
the payment handler on payment that was already paid, and the htlc
would be accepted in case this htlc was a local replay.

Now the payments database stores the payment parts details for each
incoming payments, including the htlc id and channel id for lightning
payments. We can now check that an incoming htlc on an already paid
invoice is a replay (hence accepted) or a whole new htlc (and then
rejected).