Add feature bits for custom wallet features
Wallets should explicitly advertise support for pay-to-open, trusted swaps
and wake-up notifications  in their `init` message, to allow the LSPs they
connect to to know what they're allowed to do.

This lets LSPs know for example whether they can open a channel on-the-fly
when liquidity is missing to receive a payment.