do not compare pay-to-open amount to min-funding
This is a regression introduced by #283. We should use
`PayToOpenRequest.fundingSatoshis`, not `PayToOpenRequest.amountMsat`
when comparing to `nodeParams.minFundingSatoshis`.

NB: this operates on a per-payment-part basis, and in the case of an AMP
payment, the final funding amount will be different, but it won't be
smaller than the one of each single part. If it were, because of a bug in
the peer, the channel would be rejected by Phoenix at opening and that's
okay. The peer would retry until a correct parameter is set and Phoenix
accepts the channel.

In theory we could rely on the `PayToOpenRequest.combine` method to
compute the final `fundingSatoshi`, but there is no guarantee that
Phoenix has the same implementation than the peer so it's not better.
The `combine()` wasn't used and has been removed.