Subtract fees from pay-to-open amounts in database
Remove WalletPayment companion object, and add amount/fees abstract field.
The amount always reflects the change in the wallet's balance and is set
after applying the fees:
- for outgoing payments, this the total amount that left the wallet, including
the fees.
- for incoming payments, this is the total amount that ended up in the wallet,
after applying fees if any (e.g for creating a channel)

This changes the way the pay-to-open amount is stored in the database: the
channel creation fee is now subtracted for the p2o amount. The amount is what
is pushed to us in the channel.

See https://github.com/ACINQ/phoenix/issues/170