Prune legacy recipient routing hints (#309)
When sending to a recipient that doesn't support trampoline, we provide
the invoice's routing hints to the trampoline node in the onion. If too
many routing hints are provided, this will overflow the onion.

We now limit the number of routing hints that are sent to the trampoline
node to avoid overflowing the onion

A better long term solution is to send the routing hints in a tlv in
`update_add_htlc` instead of including them in the onion.

An even better long term solution is to support trampoline, which removes
the need for these routing hints in the first place.