Schedule `checkPaymentsTimeout` every 10s (#320)
With the previous interval of 30s, a payment could stay pending for up
to 90s in the worst case scenario. This opens edge cases with the
pay-to-open logic, where expired pending parts could be taken into
account when processing other parts with the same payment hash.

cherry-picked from #62c59d6