Improve legacy Phoenix InitTLVs and messages (#325)
A new stream parameter has been added to Peer. This stream is optional and is 
empty by default. It can be used to add TLVs to the Init message sent to the 
Peer. The InitTLV.PhoenixAndroidLegacyNodeId object now contains a new
signature field to prevent leaks.

The KeyManager now provides a method to get the legacy node key as computed
by eclair-core.

Support has been added for the PhoenixAndroidLegacyInfo Lightning message 
which is replied by the peer after receiving the PhoenixAndroidLegacyNodeId TLV.
This event is then broadcast in the PeerListener channel.

Co-authored-by: Pierre-Marie Padiou <pm47@users.noreply.github.com>