Add support for swap-out payments (#342)
Custom lightning messages and peer events have been added for swap-out 
payments.

The object storing the details of a swap-out payment in the database has been
amended. It now contains a payment request field (swap-out are actually lightning
payments sent to a lightning node, that will in turn make an on-chain transaction).
The details object also contains a new field for the swap-out fee

The SendPayment class is now an interface. It can either stand for a normal 
lightning payments or a swap-out payment.