Add UUID to ReceivedWith.NewChannel to make these parts unique
Incoming payment parts are contained in a set that ensure that parts
are not duplicated (esp. the `LightningPayment` part where the tuple
of (channelId, htlcId) is unique). So the `NewChannel` part type must
also have a way to ensure unicity, since an incoming payment can be
settled with a new channel done in several pay-to-open parts. This is
done using a random UUID that is assigned when the part is processed
by the incoming payment handler.

This commit fixes an issue where the new-channel parts of an incoming
payment would be lost if those parts had the same amount.