Add UUID to ReceivedWith.NewChannel to ensure unicity (#345)
Incoming payment are received through one or several parts, which are 
contained in a `Set<ReceivedWith>`. Each part must be unique.

This works well for `LightningPayment` parts which have a unique tuple
of (channelId, htlcId), but not for `NewChannel` parts. We fix this by 
adding a new `id: UUID` field to that object and assigning that field when 
the part is processed by the incoming payment handler.

This commit fixes an issue where `NewChannel` parts with the same fee,
amount and channel would be trimmed to only one single part.