Add dual funding contributions (#352)
Add helper function that takes some utxos and will create the corresponding
dual funding contributions. The change output amount (if any) is set to
ensure that fees are correctly paid by each participant according to the
specification.

We then finalize the dual funding flow by exchanging signatures for the
shared transaction.

We now use deterministic serial IDs. We don't need them to be random,
and it adds the risk of accidentally re-using the same serial ID twice.