Work around lnd interpreting invoice feature bits
This is a work-around for a compatibility issue between lnd and our legacy
trampoline feature bit, which was using feature bits 50/51 that have been
assigned to zero-conf. We need our invoices to still contain this legacy
trampoline feature bit to allow old Phoenix wallets to pay new Phoenix
wallets using trampoline.

But lnd interprets that as zero-conf (even though it's not marked as an
invoice feature in Bolt 9) and thus requires the scid-alias feature bit (47)
to also be set.