Wait for dual funding tx confirmation
We now only put a watch-confirmed on the funding tx while it's unconfirmed,
and put a watch-spent only after we received the watch-confirmed event.

We always republish funding transactions after a restart, and remove the
funding timeout mechanism: we will reintroduce it later, but it will have
to check multiple funding tx candidates so it cannot use the previous
mechanism.

This also lets us do some clean-up of the electrum watcher component.