Add support for funding tx RBF
We add support for the RBF flow of the interactive-tx construction.
The initiator side is very basic as it is only used in tests.

Note that this requires a serialisation trick to add alternative commitments
to the closing state, which is handled in a backwards-compatible way.

We also remove some unused, overly complex code that was initially imported
from eclair.