Add codecs for please_open_channel
This is a new message that lets a (future) non-initiator ask an LSP to
open a channel (against a small fee).

This mechanism will be used to replace the previous trusted swap-in
protocol: the result will be a fully trustless protocol where the wallet
user chooses which utxos to add to the channel (thanks to dual funding).

We also include a temporary tlv that helps provide smooth migration
for Phoenix users on android.