Add please_open_channel protocol
When a wallet has on-chain funds available, they can signal their LSP with
a `please_open_channel` message. The LSP will respond with an `open_channel`
asking for a fee for the new channel. If the wallet accepts, we move on to
the standard dual funding flow where both wallet and LSP contribute.