Update Phoenix channel type (#372)
* Add zero-reserve channel type

Add a new supported channel type using zero-reserve but not zero-conf.
This is necessary when using dual funding where both parties contribute,
so there is a risk that one party cheats and double-spends the funding tx.

* Deprecate experimental zero-conf feature

We instead rely on a peer whitelist to accept zero-conf or not.

This lets us use a single channel type for Phoenix, and decide whether to
allow zero-conf more dynamically based on details of the operation.

* Send scid-alias in ChannelReady

We cannot add a tlv stream to FundingLocked without breaking the
backwards-compatibility with our previous serialization. Instead, we
replace FundingLocked with a new ChannelReady message, and keep the
old message around for the legacy states.