Add `please_open_channel` mechanism for trustless swap-in (#373)
This is a new message that lets a (future) non-initiator ask an LSP to
open a channel (against a small fee). When a wallet has on-chain funds
available, they can signal their LSP with a `please_open_channel`
message. The LSP will respond with an `open_channel` asking for a
fee for the new channel. If the wallet accepts, we move on to the
standard dual funding flow where both wallet and LSP contribute.

This mechanism will be used to replace the previous trusted swap-in
protocol: the result will be a fully trustless protocol where the wallet
user chooses which utxos to add to the channel (thanks to dual funding).

We also include a temporary tlv that helps provide smooth migration
for Phoenix users on android.

Co-authored-by: pm47 <pm.padiou@gmail.com>