Support Electrum requests multiplexing (#384)
This is the minimum change to make `ElectrumClient` support multiple independent "callers" (watcher, wallets).

Requests made by independent callers are identified with a unique id and multiplexed on the same connection. Each caller then listens to the `notifications` flow and only keeps responses for his identifiers. For convenience, this is encapsulated in an inner class `ElectrumClient.Caller` which presents the same interface as before multiplexing.

Note that subscribtion responses are currently _not_ multiplexed. They don't have an associated id and are sent to all callers.