remove non-persisted classes from binary codecs
In an attempt to minimize potential deserialization risks, we have
duplicated `ChannelState` classes with mirror classes that are supposed
to be immutable. But we actually only need to duplicate subclasses of
`ChannelStateWithCommitments`, because the other ones are not persisted.

The confusion stems for the fact that there was an ambiguity between
json and binary serialization. But they have a different scope:
- json serialization should apply to (original) subclasses of `ChannelState`,
including the ones that are not persisted
- binary serialization should apply to (mirror) subclasses of
  `ChannelStateWithCommitments`.