Add tlv to require confirmed inputs for dual funding
Add tlv to require confirmed inputs in dual funding: when that tlv is specified,
the peer must use confirmed inputs, otherwise the funding attempt will be
rejected. This ensures that we won't pay the fees for a low-feerate ancestor.

We understand the field, but don't actually check the condition.
We always use confirmed inputs for the funds that we are adding to a
channel.