add a new v4 serialization [no ci]
It uses a custom serialization, contrary to v2 and v3 which use
`kotlinx-serialization`.

The tradeoff for initially going with `kotlinx-serialization` was to
accept using a serialization mechanism that we don't fully control, in
exchange for more automation and less code. On top of that, we get json
serialization for free.

However, because channel data is so sensitive, we are not fully
confident in the serialization system (particularly w.r.t data model
migrations). In an attempt to reduce the risk of not being able to read
older data in the future, we have mirrored most of channel classes in
the `vX` packages, and we serialize those mirror classes, instead of the
actual ones.

We end up with well over a thousand lines of code for each version of
the serialization *and* the complexity inherent to
`kotlinx.serialization` with the concept of annotations, surrogate
serializers, modules, encoders and decoders, etc. On top of that, it
turns out json doesn't come for free since the variations in encoding
between json and binary spills over in various parts of the codebase in
the form of `@Contextual` annotations. In other words the tradeoff
doesn't seem to be worth it.

With custom serialization we get full control of the system, but we have
potentially more code. It turns out that it's actually not that much
(especially if we delegate relevant parts to existing bitcoin/lightning
protocol serializers) and, critically, it's systematic code with few
concepts to be aware of. In fact, there is a good chance that we end up
with less code than what we had before for each serialization version.

One drawback of custom serialization is that, since we can't rely on
reflection, and aren't using a compiler plugin like
`kotlinx.serialization`, the handling of polymorphism isn't very nice.
It's trivial to serialize child classes by prefixing with a
discriminator, but deserializing is another matter, as we have no way
to resolve a class constructor corresponding to a discriminator. As a
result, we can't factor the discriminator<->class correspondance which
means we need to be careful to keep serialization and deserialization
in sync.

Regarding json, I believe that we should still be using
`kotlinx-seriaization`, but in a different, non-intrusive way. We could
define a `json` package with serializers generated using the new
`forClass` attribute of `@Serializable`. This would free our regular
classes of all the annotation mess, and we wouldn't need `@Contextual`
anymore due to only using json. There is a catch though, as we still
need to handle backward compatibility for binary codecs relying on those
annotations. I think we can use the same `forClass` trick to duplicate
the necessary serializers in `vX`.