introduce a PersistedChannelState state
Not all `ChannelStateWithCommitments` are persisted, for example
`Offline`, `Syncing` are not.

There is a potential backward incompatibility on old serialized data for
`ErrorInformationLeak` (which I'm not tagging as
`PersistedChannelState`), but we shouldn't have any.

Note that in eclair `Closed` isn't peristed either, but I left it here
because it does make sense, and that would certainly cause backward
compat issues anyway.