Dual-funding uses txHash and verifies nSequence (#392)
* TxSignatures uses `tx_hash` instead of `tx_id`

We usually have the whole transaction available when we sign, so we use
it in the constructor to avoid getting this wrong.

* Verify replaceable nSequence

Reject `tx_add_input` with an `nSequence` set to `0xfffffffe` or
`0xffffffff`. In theory we only need one of the inputs to have an
`nSequence` below `0xfffffffe`, but it's simpler to check all of them
to be able to fail early in the protocol.