fix deserialization of LN message
Deserializers for messages `FundingCreated` and `FundingSigned` were
removed in dd6c84cb5ec84ebfec222f67270197be92af80f0 as part of
implementing dual funding.

But we now use LN protocol format in v4 serialization. This means that
we need to be able to read/write all LN messages that are persisted.
Even if those messages have been deprecated by dual funding, we may have
existing channels in the process of being created at the time when the
user upgrades Phoenix. This is the purpose of
`LegacyWaitForFundingConfirmed`/`LegacyWaitForFundingLocked`. If these
states are written back to disk, they will use v4 format, therefore we
need to be able to serialize and deserialize all LN messages that are
part of those states, which includes `FundingCreated` and `FundingSigned`.

Why wasn't this caught in tests?

Channel state tests do a serialization roundtrip at each state
transition, but we never have to serialize
`LegacyWaitForFundingConfirmed`, it is only read and then we move to the
next state. To be fair, in reality there is probably no code path where
we would need to actually serialize this state, but this seems a bit
risky to take chances for such a little change.

Static non-reg tests only deserializes data and compare it to reference
json. This PR also adds a serialization roundtrip test, which succesfully
reproduces the issue.

Why is the error in the v2 serialization whereas the bug affects v4?

Because of the try-and-fallback logic that we used for deserializing
data. All deserialization failures end up with an error in the last
codec in the list, which happens to be v2 (cc @robbiehanson). This logic
has been changed and we now explicitly try the correct serialization
based on the version discriminator.