Put back deserialization of  `FundingCreated`/`FundingSigned` (#400)
Deserializers for messages `FundingCreated` and `FundingSigned` were
removed in dd6c84cb5ec84ebfec222f67270197be92af80f0 as part of
implementing dual funding.

But we now use LN protocol format in v4 serialization. This means that
we need to be able to read/write all LN messages that are persisted.