Emit `Confirmed` event when channel reaches normal
When a new channel has been created, it is not immediately usable for
sending. Due to previously always using zero-conf, the time during which
the channel was created but not yet usable was almost instantly. But
with dual funding, we need to wait for a few confirmations, making the
"created-but-not-usable-yet" window much larger (tens of minutes).

Note that even in the zero-conf case it is cleaner to handle this
explicitly, because even if the "created-but-not-usable-yet" was very
small, it was still non-zero.