Hardware wallet-based channel funding
Add support for opening dual-funded channels using utxos
managed by a hardware wallet.

Phoenix initiates the flow by providing an xpub it gets
from the hardware wallet: we then ask electrum to fetch
utxos available on that address and initiate the dual
funding flow.

When we get to the signing step, we send an event to
Phoenix containing a PSBT that should be signed by the
hardware wallet. Phoenix asks the hardware wallet to
sign its inputs in this PSBT and sends the signed PSBT
back. We then resume the dual-funding flow and publish
the funding transaction.

Note that we include additional metadata in the PSBT
that lets the hardware wallet verify that we have a
valid signed commitment transaction *before* it signs
the funding transaction.