Include channel backup when it fits
We previously omitted the first channel backup because it risked overflowing
the `tx_signatures` message. This is an issue for users who do a swap-in,
never use the channel, and expect to be able to restore their seed on a
different device while keeping the channel open.

We now always try to include the channel backup, and simply omit it when
it doesn't fit in the message. This should work for most users since
transaction sizes are usually well below 64kB. It won't work when the swap
transaction is too large (e.g. when doing a large CoinJoin), but in that
case our node will send an `update_fee` immediately after the funding tx
confirms, which will automatically generate a post-funding backup.