Remove reserve requirement on first commitment
If the non-initiator contributes a very large amount (more than 100x what
the initiator contributes), the initiator will not initially meet the
reserve requirements.

This is ok as long as the initiator's balance contains enough funds to pay
the commit tx fees, they don't need to meet the reserve immediately.