Dual funding latest changes (#407)
* Add tlv to require confirmed inputs for dual funding

Add tlv to require confirmed inputs in dual funding: when that tlv is specified,
the peer must use confirmed inputs, otherwise the funding attempt will be
rejected. This ensures that we won't pay the fees for a low-feerate ancestor.

We understand the field, but don't actually check the condition.
We always use confirmed inputs for the funds that we are adding to a
channel.

* Add second per-commitment point to dual funding

We can send the second per-commitment point right at the beginning of the
channel opening flow, to avoid the awkwardness of having to store an
invalid value in `remoteNextCommitInfo` until we receive `channel_ready`.

* Echo incoming tx_abort

When we receive tx_abort, we echo it back to let our peer know we've seen
their tx_abort and have reset our internal state accordingly.

Spec commit:
https://github.com/lightning/bolts/pull/851/commits/07cc0edc791aff78398a48fc31ee23b45374d8d9

* Reject negative funding amount

A negative funding amount is obviously incorrect.

* Remove reserve requirement on first commitment

If the non-initiator contributes a very large amount (more than 100x what
the initiator contributes), the initiator will not initially meet the
reserve requirements.

This is ok as long as the initiator's balance contains enough funds to pay
the commit tx fees, they don't need to meet the reserve immediately.