Persist `WaitForFundingSigned` state
We now persist our channel state right after sending our `commit_sig`.
This will let us resume the signing session if we disconnect and gets rid
of the scenario where a channel is in a state where we must wait for it
to be double-spent to be able to forget it.