store restored channel immediately
When we restore a channel from a peer backup, we need to store the
channel in database as soon as possible, so we are exactly in the same
state as if the channel was restored from local db.

This applies whether we are restoring from empty or outdated local db.