Remove public channel updates support (#435)
* Remove public channel updates support

Mobile wallets aren't relaying payments, so they don't need to advertise
the fact that a channel is disabled when a disconnection happens.

* Remove ChannelAnnouncement from Normal

We only support unannounced channels, so we remove the `buried` field
and the `channelAnnouncement` field from channel state.

Even if we wished to support public channels, we wouldn't use a `buried`
boolean but rather a new `LocalFundingStatus` (e.g. `DeeplyConfirmed`).