Handle exceptions when decoding lightning messages
When we receive a lightning message from the TCP connection, we must wrap
the decoding logic in a try/catch, otherwise invalid messages will crash
the coroutine (and potentially the wallet application).

We don't need to do that when reading messages that we previously decoded
or encoded ourselves, that's why we didn't change the signature of the
`LightningMessage.decode` function.