Rework channel features (#448)
* remove unsupported channel types

Some channel types are "understood", but not supported. This creates two code paths where we end up rejecting channel types.

Instead, we rely on `UnsupportedChannelTypes`, which is simpler and type checked.

Also, remove the `ChannelFeatures.channelType` function, which is wrong
on two counts:
- It doesn't include permanent features that are not present in the
channel type. This will be fixed in the next commit.
- The resolution is not type checked, and simply defaults to the (deleted)
`Standard` channel type.

* proper type for `Init.features`

* enrich ChannelFeatures with init features

Instead of only using features from the `ChannelType`. This makes
`DualFunding` a permanent channel feature.

* only put initFeatures in localParams.features

* only ignore commit_sig for dual-funded channels

Legacy single-funding unconfirmed channels will have `localFundingStatus=UnconfirmedFundingTx`,
so we cannot just check on the absence of `signedTx`. This is quite the
corner case, but still.