Remove `ErrorInformationLeak` state and fix channel recovery (#454)
Splicing makes the handling of multi devices more difficult, because an unknown spending tx could in fact be a splice tx. There was no real utility to the `InformationLeak` state so we remove it entirely.

There is another subtelty: when we recover from an outdated state, we also need to update the watches, because the recovered state may be one or several splices ahead. The easiest way to do that is to recreate the channel, just like if we recovered from zero.

This leads to a nice refactoring of the recovery process, where both cases (recover from scratch / recover from outdated) happen at the same place, in the Peer.

Co-authored-by: t-bast <bastien@acinq.fr>