Fix computation of fee in cpfp feerate estimator (#462)
If user attempts a cpfp with a smaller target feerate than the feerate
of the existing ancestors, the projected feerate would be negative. This
is why we correct it by taking the max between the requested feerate and
the projecte feerate.

But we forgot to do that for the absolute fee value.