make liquidity policy mutable
It is now defined as a `MutableStateFlow` and can be modified at will by the library user.

For example Phoenix users can modify their max fee setting without restarting the app and the new value will be automatically taken into account.

Note that for swap-in, we currently only do one try at startup, but that's a separate concern.