handle forward compat when restoring
In the corner case where we receive a data from the peer that has been
generated with a more recent version of lightning-kmp, we do not want
to return an error to the peer, because they will force-close and we
will be unable to do anything as we can't read the data.

Best thing is to not answer, and tell the user to upgrade the app.