Update wallet to use 2-of-2 swap-in
Phoenix uses a swap-in mechanism that allows us to make 0-conf channels
and splices. The user sends funds to an address that has two script paths:
one path that requires signatures from Phoenix and its LSP, and another
path that lets Phoenix recover the funds after a relative timeout.

We use a static public key for swaps, to make recovery simpler.