Relax reserve requirements on HTLC receiver (#463)
Bolt 2 requires the receiver of an HTLC to pay able to pay the commit tx
fee while maintaining its channel balance. This is an issue because it
can lead to a situation where the peer that has most of the channel's
funds is unable to send outgoing payments: the channel is stuck.

This becomes more frequent with splicing, which makes the channel reserve
dynamic. We thus allow the receiver to temporarily dip into its channel
reserve to pay the commit tx fee. This is ok because either:

- the HTLC is failed and removed from the commit tx, which restores the
  previous balance
- the HTLC is fulfilled and increases the receiver's balance, which helps
  them maintain their channel balance