Skip waiting for reestablish for unsigned channels
When we are a backup client, we wait for our peer to send their
`channel_reestablish` first to make sure we haven't lost data.

This can lead to a deadlock if we're waiting for a `channel_reestablish`,
but our peer has forgotten that channel because we got disconnected
before they were able to send their signatures.

To avoid this deadlock, we send our `channel_reestablish` immediately for
channels where we're waiting for signatures.