Prevent deadlocks in Peer (#484)
We were using BUFFERED channels, with the default behavior where the
sender suspends when the buffer is full (64 elements by default).

The issue is that the same coroutine receives from the `input` channel
and emits potentially multiple events to that same channel. We would thus
fill the buffer, which causes us to `suspend`, which stops dequeuing
events, leading to a complete deadlock.