Introduce a `SwapInManager`
We create a new `SwapInManager` that checks the wallet state and decides
whether to initiate a channel funding attempt, while keeping track of
utxos that are currently being used.

We allow unlocking those utxos once a channel funding attempt fails,
which ensures that if we retry later (because another input is confirmed)
we will reuse the previously failed inputs (instead of waiting for a
restart to use them).